requirements.txt
flask
google-generativeai
gunicorn
langchain
langchain-google-genai
langchain-community
faiss-cpu